+++ lab2/prof_sources/simple_cnn/image.h
@@ -27,7 +27,7 @@
  */
 void normalize_image(const unsigned char *rgb_image, float *norm_image);
 
-void image_to_ip(const float *norm_image, int8_t *image_ip);
+void image_to_ip(const float *norm_image, int16_t *image_ip);
 
 /**
  * Prints the input image to stdout in ppm format (see http://paulbourke.net/dataformats/ppm/).
